requirements.txt
openai==1.58.1
python-dotenv==1.1.1
streamlit==1.41.1
PyMuPDF==1.26.3